backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
praw>=7.7.1
nltk>=3.8.1
pandas>=2.2.0
apscheduler>=3.10.4
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.25.2
pyyaml>=6.0.1
python-dotenv>=1.0.0